refactor(provider): split ModelCatalog god-file by concern (1202->171 LOC)
Split ModelCatalog.swift into focused files: catalog data types stay in
ModelCatalog.swift; ModelCatalogClient (coordinator catalog HTTP client),
ModelDownloader core, and the ModelPrefetcher abstraction move to their own
files. ModelDownloader's interleaved methods split into +Download (manifest/
legacy orchestration), +Prefetch (resume-aware prefetch), and +HTTP (file
fetch/stream/hash/publish) extensions.

Bodies byte-identical; ModelDownloader's 4 stored props + cross-file private
methods and CatalogResponse widened private->internal. Behavior unchanged.

Build green; swift test 1052/73 pass (unchanged from baseline).
