+++ provider-swift/Sources/ProviderCore/Models/ModelCatalogClient.swift
@@ -1,0 +1,111 @@
+/// ModelCatalogClient -- coordinator-side catalog HTTP client.
+///
+/// Fetches the canonical catalog + per-model manifests from the
+/// coordinator (`GET /v1/models/catalog`, `/v1/models/{id}/manifest`).
+/// Split out of ModelCatalog.swift; see also ModelDownloader.swift.
+
+import Foundation
+
+// MARK: - Catalog client
+
+public struct ModelCatalogClient: Sendable {
+
+    private let coordinatorURL: String
+    private let urlSession: URLSession
+
+    public init(coordinatorURL: String, urlSession: URLSession = .shared) {
+        self.coordinatorURL = coordinatorHTTPBase(coordinatorURL)
+        self.urlSession = urlSession
+    }
+
+    /// Fetch the active catalog from the coordinator. `typeFilter` mirrors
+    /// the coordinator's `?type=` query parameter (e.g. "text").
+    public func fetchCatalog(typeFilter: String? = nil) async throws -> [CatalogModel] {
+        try await fetchCatalogSnapshot(typeFilter: typeFilter, includeAliases: false).models
+    }
+
+    public func fetchCatalogSnapshot(typeFilter: String? = nil, includeAliases: Bool = false) async throws -> CatalogSnapshot {
+        var components = URLComponents(string: "\(coordinatorURL)/v1/models/catalog")!
+        var queryItems: [URLQueryItem] = []
+        if let typeFilter, !typeFilter.isEmpty {
+            queryItems.append(URLQueryItem(name: "type", value: typeFilter))
+        }
+        if includeAliases {
+            queryItems.append(URLQueryItem(name: "include_aliases", value: "1"))
+        }
+        components.queryItems = queryItems.isEmpty ? nil : queryItems
+        guard let url = components.url else {
+            throw ModelCatalogError.unreachable("invalid catalog URL")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        request.timeoutInterval = 10
+        request.setValue("application/json", forHTTPHeaderField: "Accept")
+
+        let data: Data
+        let response: URLResponse
+        do {
+            (data, response) = try await urlSession.data(for: request)
+        } catch {
+            throw ModelCatalogError.unreachable(error.localizedDescription)
+        }
+
+        if let http = response as? HTTPURLResponse, !(200..<300).contains(http.statusCode) {
+            throw ModelCatalogError.http(http.statusCode, String(data: data, encoding: .utf8) ?? "")
+        }
+
+        do {
+            let decoded = try JSONDecoder().decode(CatalogResponse.self, from: data)
+            return CatalogSnapshot(models: decoded.models, aliases: decoded.aliases ?? [])
+        } catch {
+            throw ModelCatalogError.decodeFailed(error.localizedDescription)
+        }
+    }
+
+    /// Fetch the active registry manifest for a model. Model IDs can contain
+    /// `/`, so the ID is percent-encoded as one path suffix.
+    public func fetchManifest(modelID: String) async throws -> ModelManifest {
+        guard let escapedID = Self.escapeModelIDForPath(modelID),
+              let url = URL(string: "\(coordinatorURL)/v1/models/catalog/manifest/\(escapedID)")
+        else {
+            throw ModelCatalogError.unreachable("invalid manifest URL")
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "GET"
+        request.timeoutInterval = 10
+        request.setValue("application/json", forHTTPHeaderField: "Accept")
+
+        let data: Data
+        let response: URLResponse
+        do {
+            (data, response) = try await urlSession.data(for: request)
+        } catch {
+            throw ModelCatalogError.unreachable(error.localizedDescription)
+        }
+
+        if let http = response as? HTTPURLResponse, !(200..<300).contains(http.statusCode) {
+            throw ModelCatalogError.http(http.statusCode, String(data: data, encoding: .utf8) ?? "")
+        }
+
+        do {
+            return try Self.manifestDecoder.decode(ModelManifest.self, from: data)
+        } catch {
+            throw ModelCatalogError.decodeFailed(error.localizedDescription)
+        }
+    }
+
+    static let manifestDecoder: JSONDecoder = {
+        let decoder = JSONDecoder()
+        decoder.dateDecodingStrategy = .iso8601
+        return decoder
+    }()
+
+    private static func escapeModelIDForPath(_ modelID: String) -> String? {
+        var allowed = CharacterSet.urlPathAllowed
+        allowed.remove(charactersIn: "/")
+        return modelID.addingPercentEncoding(withAllowedCharacters: allowed)
+    }
+}
+
